+++ CHANGES.md
@@ -37,6 +37,7 @@ Change Log
 * Allowed Bing Maps servers with a subpath (instead of being at the root) to work correctly. [#6597](https://github.com/AnalyticalGraphicsInc/cesium/pull/6597)
 * Fixed polygon outline when using `perPositionHeight` and `extrudedHeight`. [#6595](https://github.com/AnalyticalGraphicsInc/cesium/issues/6595)
 * Fixed broken links in documentation of `createTileMapServiceImageryProvider`. [#5818](https://github.com/AnalyticalGraphicsInc/cesium/issues/5818)
+* Fixed bug where child tiles with transforms were not being clipped correctly. [#6600](https://github.com/AnalyticalGraphicsInc/cesium/issues/6600)
 
 ### 1.45 - 2018-05-01